… pkg/commands/git_commands/patch_test.go
… pkg/commands/git_commands/rebase_test.go